SPLAT-1442 change casting flow to render properly on FE
